moved nav logic to metaData
